Rework DNS request locking
Comparing tests using single vs. multiple workers showed how using
more than one worker resulted in more UDP queries being sent than when
using a single worker. The reason for this is that multiple workers can
investigate NS or MX records pointing to the same names at the same
time. In this case the previous "try to load result from cache,
otherwise do lookup and store result" operation is not enough to
guarantee there will be only a single lookup for the same NS or MX name.

Now the code has been reworked so we actually do locking around cachable
AAAA records (e.g. those being looked up because either a NS or MX
record points to them). While this of course leads to more blocking
of workers it seems reasonable to optimize for sending (and waiting on)
fewer network requests.

It is still possible for the tool to generate duplicate lookups for the
"www" subdomain. This happens if a zone also has NS or MX records
pointing to "www.<zone name>". This is a tradeoff where I assume that
doing so is fairly uncommon and it saves us from having to cache every
single "www" lookup which is done for every name in the zone.