Invert "valid a/aaaa response" checks
Now we only require a single A or AAAA record in the answer section to
be happy with the response rather than requiring only the presence of A
or AAAA. This means there can be other "strange" records in the answer
section but this is ignored.

Also simplify away the netip checks since if miekg/dns is happy with the
response it means the rdata already contained the expected number of
bytes for either a v4 or v6 address.