Make it possible to keep count of seen MX targets
Running the tool with '-mx-suffixes .mail.example.net' will check for MX
records pointing to a name that has that suffix, e.g.
"server1.mail.example.net", and if it does it will increment a counter
for it.

The final counts will be presented under the mx_suffix_counter key if
any match is found.