Improve MX suffix counting
Before this we would increase the counter any time we saw a matching
record in an RRSet, meaning that if a zone returned multiple MX records
where none of them had AAAA records we would count all of them. Ideally
the counters should not reflect the individual number of times we have
seen matching records, but rather how many zones we have seen that has
one or more matching records in them.

To fix this we need to change some things:
* We create a mxSuffixTracker map to keep track of what suffixes we have
  already seen. We only increment the counter the first time a suffix
  is matched for a given name.
* The code would stop iterating over the RRSet after the first record that
  resulted in a valid AAAA lookup. This was good before, but since we
  now possibly have multiple interesting MX records in the set we need
  to keep iterating even if a match has been found. In this case we will
  do no further DNS lookups, we just iterate locally over the remaining
  entries.